Adding PTS reader
- also fixing overflow in progress bar when loading more than 20 million points

+++ src/nz/ac/aut/SentienceLab/PointCloudDatasetReader/PointCloudData.java
@@ -38,7 +38,7 @@ public class PointCloudData
     }
     
     
-    public int pointCount()
+    public long pointCount()
     {
         return points.size();
     }
